webApiRestFul/requirements.txt
# FastAPI y servidor ASGI
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Base de datos Oracle
oracledb==2.0.1

# Autenticación y seguridad
PyJWT==2.8.0
python-multipart==0.0.9

# Validación de datos
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0

# Utilidades
python-dotenv==1.0.1
